feat(energy): only sites

+++ frontend/src/views/energy/EnergyHome.vue
@@ -7,75 +7,34 @@
     </v-row>
     <v-row>
       <v-col cols="6">
-        <v-expansion-panels multiple>
+        <v-expansion-panels mandatory multiple>
           <v-expansion-panel v-for="group in groups" :key="group.name">
             <v-expansion-panel-header>
-              {{ group.name }}
+              <h3>{{ group.name }} {{ group.emoji }}</h3>
             </v-expansion-panel-header>
             <v-expansion-panel-content>
-              <h2>
-                Templates
-                <v-btn color="primary" icon @click="createTemplate(group.name)">
-                  <v-icon>mdi-plus-box</v-icon>
-                </v-btn>
-              </h2>
-              <v-list nav>
-                <v-list-item-group>
-                  <v-list-item
-                    v-for="item in group.templates"
-                    :key="item._id"
-                    @click="openTemplate(item)"
-                  >
-                    <v-list-item-content>
-                      <v-list-item-title>
-                        {{ getName(item) }}
-                      </v-list-item-title>
-                    </v-list-item-content>
-                    <v-list-item-action class="flex-row">
-                      <v-tooltip bottom>
-                        <template v-slot:activator="{ on, attrs }">
-                          <v-btn
-                            v-bind="attrs"
-                            color="primary"
-                            icon
-                            v-on="on"
-                            @click="
-                              $event.stopPropagation();
-                              createSite(item);
-                            "
-                          >
-                            <v-icon>mdi-plus-box</v-icon>
-                          </v-btn>
-                        </template>
-                        <span>New site from template</span>
-                      </v-tooltip>
-                      <v-btn
-                        icon
-                        @click="
-                          $event.stopPropagation();
-                          deleteTemplate(item);
-                        "
-                      >
-                        <v-icon>mdi-close</v-icon>
-                      </v-btn>
-                    </v-list-item-action>
-                  </v-list-item>
-                </v-list-item-group>
-              </v-list>
-              <h2>Sites</h2>
               <v-list nav>
                 <v-list-item-group>
                   <v-list-item
                     v-for="item in group.sites"
                     :key="item._id"
-                    @click="openSite(item)"
+                    @click="openSite(item._id)"
                   >
                     <v-list-item-content>
                       <v-list-item-title>
-                        {{ getName(item) }}
+                        {{ item.name }}
                       </v-list-item-title>
                     </v-list-item-content>
-                    <v-list-item-action>
+                    <v-list-item-action class="flex-row">
+                      <v-btn
+                        icon
+                        @click="
+                          $event.stopPropagation();
+                          copySite(item);
+                        "
+                      >
+                        <v-icon>mdi-content-copy</v-icon>
+                      </v-btn>
                       <v-btn
                         icon
                         @click="
@@ -93,12 +52,23 @@
           </v-expansion-panel>
         </v-expansion-panels>
         <br />
-        <v-text-field
-          v-model="newTemplateName"
-          label="Name"
-          append-outer-icon="mdi-plus-box"
-          @click:append-outer="addNewTemplate"
-        ></v-text-field>
+        <v-card>
+          <v-form v-model="formValid" @submit.prevent="createSite">
+            <v-card-text>
+              <form-item-component
+                v-for="item in formItems"
+                :key="item.key"
+                v-model="newSiteProp[item.key]"
+                v-bind="item"
+              ></form-item-component>
+            </v-card-text>
+            <v-card-actions>
+              <v-btn color="primary" :disabled="!formValid" icon type="submit">
+                <v-icon>mdi-plus-box</v-icon>
+              </v-btn>
+            </v-card-actions>
+          </v-form>
+        </v-card>
       </v-col>
       <v-col cols="6">
         <energy-map aspect-ratio="1" :markers="markers"></energy-map>
@@ -107,65 +77,78 @@
     <v-overlay :value="loading">
       <v-progress-circular indeterminate size="64"></v-progress-circular>
     </v-overlay>
+    <confirm-dialog ref="confirmDialog"></confirm-dialog>
   </v-container>
 </template>
 
 <script lang="ts">
+import ConfirmDialog from "@/components/commons/ConfirmDialog.vue";
+import FormItemComponent, {
+  FormItem,
+} from "@/components/commons/FormItemComponent.vue";
 import EnergyMap from "@/components/energy/EnergyMap.vue";
 import { ExistingDocument } from "@/models/couchdbModel";
 import { GeneralModule, ProjectDocument } from "@/models/energyModel";
 import { energy } from "@/utils/apps";
 import { SyncDatabase } from "@/utils/couchdb";
+import getFlagEmoji from "@/utils/flagEmoji";
 import { LatLngExpression } from "leaflet";
 import { chain, cloneDeep, groupBy } from "lodash";
 import "vue-class-component/hooks";
-import { Component, Vue } from "vue-property-decorator";
+import { Component, Ref, Vue } from "vue-property-decorator";
 import { mapState } from "vuex";
 
 @Component({
   components: {
+    ConfirmDialog,
     EnergyMap,
+    FormItemComponent,
   },
   computed: {
-    ...mapState("energy", [
-      "sites",
-      "sitesDatabase",
-      "templates",
-      "templatesDatabase",
-    ]),
+    ...mapState("energy", ["sites", "sitesDatabase"]),
   },
 })
 export default class EnergyHome extends Vue {
   sites!: ExistingDocument<ProjectDocument>[];
   sitesDatabase!: SyncDatabase<ProjectDocument>;
-  templates!: ExistingDocument<ProjectDocument>[];
-  templatesDatabase!: SyncDatabase<ProjectDocument>;
   title = energy.title;
   loading = false;
-  newTemplateName = "";
+  formValid = false;
+  newSiteProp: NewSiteProp = {};
+
+  get formItems(): FormItem<keyof NewSiteProp>[] {
+    return [
+      {
+        type: "text",
+        key: "projectName",
+        label: "Project Name",
+      },
+    ];
+  }
+
+  get username(): string {
+    return this.$userName();
+  }
 
   get groups(): Group[] {
-    const groupedTemplates = groupBy(
-      this.templates,
-      (template) => template.name
-    );
-    const groupedSites = groupBy(this.sites, (site) => site.name);
-    const names: string[] = chain([
-      ...Object.keys(groupedTemplates),
-      ...Object.keys(groupedSites),
-    ])
-      .uniq()
-      .sort()
+    return chain(
+      groupBy(this.sites, (site) => site.modules.general?.name ?? "No Site")
+    )
+      .entries()
+      .sortBy(([key]) => key)
+      .map(([key, value]) => {
+        const countryCode = value[0].modules.general?.countryCode ?? "";
+        return {
+          name: key,
+          emoji: getFlagEmoji(countryCode),
+          sites: value,
+        };
+      })
       .value();
-    return names.map((name) => ({
-      name: name,
-      templates: groupedTemplates[name] ?? [],
-      sites: groupedSites[name] ?? [],
-    }));
   }
 
   get markers(): LatLngExpression[] {
-    return chain([...this.sites, ...this.templates])
+    return chain(this.sites)
       .map((site) => site.modules.general)
       .filter((general): general is GeneralModule => !!general)
       .map(
@@ -179,46 +162,45 @@ export default class EnergyHome extends Vue {
       .value();
   }
 
-  getName(document: ExistingDocument<ProjectDocument>): string {
-    return document.modules.general?.name ?? document.name;
+  @Ref()
+  readonly confirmDialog!: ConfirmDialog;
+
+  openSite(documentId: string): void {
+    this.$router.push({ path: `sites/${documentId}`, append: true });
   }
 
-  openTemplate(document: ExistingDocument<ProjectDocument>): void {
-    this.$router.push({ path: `templates/${document._id}`, append: true });
-  }
-
-  openSite(document: ExistingDocument<ProjectDocument>): void {
-    this.$router.push({ path: `sites/${document._id}`, append: true });
-  }
-
-  createTemplate(name: string): void {
-    const username = this.$userName();
+  createSite(): void {
     this.applyAsync(
-      this.templatesDatabase.remoteDB.post({
-        name: name,
-        users: [username],
-        modules: {},
-      })
+      this.sitesDatabase.remoteDB
+        .post({
+          name: this.newSiteProp.projectName ?? "",
+          users: [this.username],
+          modules: {},
+        })
+        .then((response) => {
+          this.openSite(response.id);
+        })
     );
   }
 
-  createSite(templateDocument: ExistingDocument<ProjectDocument>): void {
-    const username = this.$userName();
+  copySite(document: ExistingDocument<ProjectDocument>): void {
     this.applyAsync(
       this.sitesDatabase.remoteDB.post({
-        name: templateDocument.name,
-        users: [username],
-        modules: cloneDeep(templateDocument.modules),
+        name: `${document.name} - copy`,
+        users: [this.username],
+        modules: cloneDeep(document.modules),
       })
     );
-  }
-
-  deleteTemplate(document: ExistingDocument<ProjectDocument>): void {
-    this.applyAsync(this.templatesDatabase.remoteDB.remove(document));
   }
 
   deleteSite(document: ExistingDocument<ProjectDocument>): void {
-    this.applyAsync(this.sitesDatabase.remoteDB.remove(document));
+    this.confirmDialog
+      .open(`Are you sure to delete "${document.name}" ?`)
+      .then((confirmed) => {
+        if (confirmed) {
+          this.applyAsync(this.sitesDatabase.remoteDB.remove(document));
+        }
+      });
   }
 
   applyAsync<T>(promise: Promise<T>): void {
@@ -230,18 +212,15 @@ export default class EnergyHome extends Vue {
       })
       .finally(() => (this.loading = false));
   }
-
-  addNewTemplate(): void {
-    if (this.newTemplateName) {
-      this.createTemplate(this.newTemplateName);
-      this.newTemplateName = "";
-    }
-  }
 }
 
 interface Group {
   name: string;
-  templates: ExistingDocument<ProjectDocument>[];
+  emoji: string;
   sites: ExistingDocument<ProjectDocument>[];
+}
+
+interface NewSiteProp {
+  projectName?: string;
 }
 </script>
